move paraphrasing files to a separate folder
